+++ src/engine/relays/utils.ts
@@ -173,6 +173,10 @@ export const getUserHints = hintSelector(function* (mode: RelayMode) {
 export const getUserHint = (pubkey: string): string => first(getUserHints(1, "write")) || ""
 
 export const getEventHints = hintSelector(function* (event: Event) {
+  for (const address of Tags.from(event).circles().all()) {
+    yield* getGroupHints(address)
+  }
+
   yield* getPubkeyRelayUrls(event.pubkey, RelayMode.Write)
   yield* event.seen_on.filter(isShareableRelay)
 })
@@ -183,6 +187,10 @@ export const getEventHint = (event: Event) => first(getEventHints.limit(1).getHi
 // advertised would be the most reliable option, since well-behaved clients
 // will write replies there.
 export const getReplyHints = hintSelector(function* (event) {
+  for (const address of Tags.from(event).circles().all()) {
+    yield* getGroupHints(address)
+  }
+
   yield* getPubkeyRelayUrls(event.pubkey, RelayMode.Read)
 })
 
@@ -203,6 +211,10 @@ export const getRootHints = hintSelector(function* (event) {
 // relays. Limit how many per pubkey we publish to though. We also want to advertise
 // our content to our followers, so publish to our write relays as well.
 export const getPublishHints = hintSelector(function* (event: Event) {
+  for (const address of Tags.from(event).circles().all()) {
+    yield* getGroupHints(address)
+  }
+
   const pubkeys = Tags.from(event).type("p").values().all()
   const hintGroups = pubkeys.map(pubkey => getPubkeyRelayUrls(pubkey, RelayMode.Read))
   const authorRelays = getPubkeyRelayUrls(event.pubkey, RelayMode.Write)