+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 - [x] Show all replies exands notes inline rather than opening the note
 - [x] Fix nip04 messages not having a members field
 - [x] Improve person search
+- [x] Fix zaps in private groups
+- [x] Improve relay selection for groups
 
 # 0.4.0